requirements.txt
aiohttp>=3.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
